Added lang vars Split afc into sub components
